Log line matchers: consolidate request date pattern matching
Bouncer can have log messages interleaved in an interesting way:
A log message emitted by the web application logging facility
is emitted through Python's StreamHandler which calls write()
once for the message, and write() another time for the trailing
newline. The logging facility of gunicorn's worker process
(or master process?) is not coordinated with that of the web
application's logging facility, which is why log lines emitted
by them can interleave. It's most likely for them to interleave
exactly inbetween the byte sequences emitted by individual
write() system calls. Which is why we saw this log line:

Jul 02 11:34:01 saber-dcos-dev-master-az1-192-168-10-20.novalocal bouncer.sh[24589]: [180702-11:34:01.456] [24613:Thread-6] [bouncer.app.internal.PolicyQuery] INFO: type=audit timestamp=2018-07-02T16:34:01.456876Z srcip=127.0.0.1 authorizer=bouncer uid=as033901 action=full object=dcos:adminrouter:ops:slave result=allow reason="user group `default` in ACL, action `full` set"192.168.10.22 [02/Jul/2018:11:34:01 -0500] "GET /acs/api/v1/internal/policyquery?rid=dcos:adminrouter:ops:mesos&uid=dcos_history_service&action=full HTTP/1.0" 200 22 "-" "-" (0.214849 s)

It starts with a systemd journald prefix and continues with
a bouncer web application-emitted log line which is then *not*
terminated, but immediately followed by a log message emitted
by gunicorn (this time not prefixed with the journald prefix,
because it was not recognized as individual log message).

This is the portion where a newline byte was missing:

        `full` set"192.168

(between " and 1).

The current regex for matching the request date then matches a
large part of the entire intertwined/combined log message,
and parsing the date from that matching group fails (obviously).

Use the 'Negated Class Solution' for consolidating the part of
the pattern that is supposed to match the request date only.

The 'Negated Class Solution' makes use of the fact that we know
that the character [ will never be present between the start
delimiter [ and the end delimiter ].

With this method we could also git rid of the lazy quantifier
(?), but I am not sure if this is might reduce performance.